requirements.txt
agno==1.7.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.13.0
dataclasses-json==0.6.7
dateparser==1.2.2
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gritql==0.2.0
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-cli==0.0.36
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-ollama==0.3.6
langchain-openai==0.3.28
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.9
langgraph==0.5.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langserve==0.3.1
langsmith==0.4.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.12
mypy_extensions==1.1.0
networkx==3.5
numexpr==2.11.0
numpy==2.3.1
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.9.79
oauthlib==3.3.1
ollama==0.5.1
openai==1.97.1
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-triton==3.4.0+git11ec6354
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==1.8.2
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.9.0.dev20250727+cu129
torchaudio==2.8.0.dev20250727+cu129
torchvision==0.24.0.dev20250727+cu129
tqdm==4.67.1
transformers==4.54.0
triton==3.3.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
yfinance==0.2.65
zstandard==0.23.0
